+++ packages/agw-client/src/actions/signTransaction.ts
@@ -6,6 +6,7 @@ import {
   encodeAbiParameters,
   type Hex,
   parseAbiParameters,
+  type PublicClient,
   type Transport,
   type UnionRequiredBy,
   type WalletClient,
@@ -25,6 +26,7 @@ import {
   type AssertEip712RequestParameters,
 } from '../eip712.js';
 import { AccountNotFoundError } from '../errors/account.js';
+import { assertSessionKeyPolicies } from '../sessionValidator.js';
 import type { CustomPaymasterHandler } from '../types/customPaymaster.js';
 import { VALID_CHAINS } from '../utils.js';
 import { transformHexValues } from '../utils.js';
@@ -37,6 +39,7 @@ export async function signTransaction<
 >(
   client: Client<Transport, ChainEIP712, Account>,
   signerClient: WalletClient<Transport, ChainEIP712, Account>,
+  publicClient: PublicClient<Transport, ChainEIP712>,
   args: SignEip712TransactionParameters<chain, account, chainOverride>,
   validator: Address,
   useSignerAddress = false,
@@ -56,6 +59,7 @@ export async function signTransaction<
   const { transaction, customSignature } = await signEip712TransactionInternal(
     client,
     signerClient,
+    publicClient,
     args,
     validator,
     useSignerAddress,
@@ -80,6 +84,7 @@ export async function signEip712TransactionInternal<
 >(
   client: Client<Transport, ChainEIP712, Account>,
   signerClient: WalletClient<Transport, ChainEIP712, Account>,
+  publicClient: PublicClient<Transport, ChainEIP712>,
   args: SignEip712TransactionParameters<chain, account, chainOverride>,
   validator: Address,
   useSignerAddress = false,
@@ -104,7 +109,6 @@ export async function signEip712TransactionInternal<
     'maxFeePerGas',
     'maxPriorityFeePerGas',
     'gas',
-    'value',
     'chainId',
     'gasPerPubdata',
   ]);
@@ -136,12 +140,25 @@ export async function signEip712TransactionInternal<
       chain: chain,
     });
 
+  await assertSessionKeyPolicies(
+    publicClient,
+    chainId,
+    fromAccount,
+    transaction,
+  );
+
   const transactionWithPaymaster = await getTransactionWithPaymasterData(
     chainId,
     fromAccount,
     transaction,
     customPaymasterHandler,
   );
+
+  if (transactionWithPaymaster.data === undefined) {
+    // serializer turns undefined into 0x00 which causes issues sending
+    // eth to contracts that don't have a fallback function
+    transactionWithPaymaster.data = '0x';
+  }
 
   const eip712Domain = chain?.custom.getEip712Domain({
     ...transactionWithPaymaster,