+++ packages/agw-client/src/walletActions.ts
@@ -5,6 +5,7 @@ import {
   type Chain,
   type Client,
   type GetChainIdReturnType,
+  type Hash,
   type PrepareTransactionRequestReturnType,
   type PublicClient,
   type SendTransactionRequest,
@@ -47,6 +48,7 @@ import {
   isLinkedAccount,
   type IsLinkedAccountParameters,
 } from './actions/getLinkedAgw.js';
+import { getSessionStatus } from './actions/getSessionStatus.js';
 import {
   linkToAgw,
   type LinkToAgwParameters,
@@ -78,7 +80,7 @@ import { writeContract } from './actions/writeContract.js';
 import { writeContractForSession } from './actions/writeContractForSession.js';
 import { EOA_VALIDATOR_ADDRESS } from './constants.js';
 import { type SessionClient, toSessionClient } from './sessionClient.js';
-import type { SessionConfig } from './sessions.js';
+import type { SessionConfig, SessionStatus } from './sessions.js';
 import type { CustomPaymasterHandler } from './types/customPaymaster.js';
 import type { SendTransactionBatchParameters } from './types/sendTransactionBatch.js';
 
@@ -127,6 +129,9 @@ export type AbstractWalletActions<
     >,
   ) => Promise<PrepareTransactionRequestReturnType>;
   toSessionClient: (signer: Account, session: SessionConfig) => SessionClient;
+  getSessionStatus: (
+    sessionHashOrConfig: Hash | SessionConfig,
+  ) => Promise<SessionStatus>;
 };
 
 // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
@@ -151,6 +156,7 @@ export type SessionClientActions<
   ) => Promise<SignTransactionReturnType>;
   writeContract: WalletActions<chain, account>['writeContract'];
   signTypedData: WalletActions<chain, account>['signTypedData'];
+  getSessionStatus: () => Promise<SessionStatus>;
 };
 
 export type LinkableWalletActions<
@@ -210,9 +216,16 @@ export function sessionWalletActions(
       signTypedDataForSession(
         client,
         signerClient,
+        publicClient,
         args as any,
         session,
         paymasterHandler,
+      ),
+    getSessionStatus: () =>
+      getSessionStatus(
+        publicClient,
+        parseAccount(client.account).address,
+        session,
       ),
   });
 }
@@ -268,6 +281,7 @@ export function globalWalletActions<
       signTransaction(
         client,
         signerClient,
+        publicClient,
         args as SignEip712TransactionParameters<chain, account>,
         EOA_VALIDATOR_ADDRESS,
         false,
@@ -277,7 +291,8 @@ export function globalWalletActions<
       ),
     signTypedData: (
       args: Omit<SignTypedDataParameters, 'account' | 'privateKey'>,
-    ) => signTypedData(client, signerClient, args, isPrivyCrossApp),
+    ) =>
+      signTypedData(client, signerClient, publicClient, args, isPrivyCrossApp),
     deployContract: (args) =>
       deployContract(client, signerClient, publicClient, args, isPrivyCrossApp),
     writeContract: (args) =>
@@ -313,6 +328,12 @@ export function globalWalletActions<
         session: session,
         paymasterHandler: customPaymasterHandler,
       }),
+    getSessionStatus: (sessionHashOrConfig: Hash | SessionConfig) =>
+      getSessionStatus(
+        publicClient,
+        parseAccount(client.account).address,
+        sessionHashOrConfig,
+      ),
   });
 }