+++ packages/agw-client/test/src/walletActions.test.ts
@@ -92,6 +92,7 @@ describe('globalWalletActions', () => {
     expect(signTransactionModule.signTransaction).toHaveBeenCalledWith(
       mockClient,
       mockSignerClient,
+      mockPublicClient,
       mockArgs,
       EOA_VALIDATOR_ADDRESS,
       false,